requirements.txt
langchain==0.2.1
langchain-core==0.2.3
langchain-community==0.2.1
langchain-google-genai==1.0.4
langchain-groq==0.1.6
langsmith==0.1.77
pydantic<3,>=1
nest_asyncio
sentence-transformers
qdrant-client
openai
python-dotenv
PyPDF2
pdf2image
pytesseract
streamlit
streamlit_mic_recorder
pyttsx3
accelerate
google-api-python-client
beautifulsoup4
duckduckgo-search
reportlab
